Testing time zone cast

+++ src/hotspot/share/runtime/os.cpp
@@ -195,6 +195,9 @@ char* os::iso8601_time(char* buffer, size_t buffer_length, bool utc) {
   // Print an ISO 8601 date and time stamp into the buffer
   const int year = 1900 + time_struct.tm_year;
   const int month = 1 + time_struct.tm_mon;
+  // Cast zone_hours and zone_min to int to avoid misalignment in variadic
+  // argument passing on platforms where time_t is wider than int (e.g.,
+  // 32-bit ARM with musl libc which uses 64-bit time_t for Y2038 safety).
   const int printed = jio_snprintf(buffer, buffer_length,
                                    "%04d-%02d-%02dT%02d:%02d:%02d.%03d%c%02d%02d",
                                    year,
@@ -205,8 +208,8 @@ char* os::iso8601_time(char* buffer, size_t buffer_length, bool utc) {
                                    time_struct.tm_sec,
                                    milliseconds_after_second,
                                    sign_local_to_UTC,
-                                   zone_hours,
-                                   zone_min);
+                                   (int)zone_hours,
+                                   (int)zone_min);
   if (printed == 0) {
     assert(false, "Failed jio_printf");
     return NULL;
